test: Get test code building under Windows
 - Initial nmake file based off library nmake file
 - Cast socket call arguments to (char *)
 - Use wrapper sclose() that maps to closesocket() or close()
 - Build a config.h indicating presence of headers
 - Conditionally include netdb.h
 - Remove unnecessary include of sys/socket.h
 - Force longer bitmask for allocation failure tracking
 - Call WSAStartup() / WSACleanup() in main()
 - Set TCP_NODELAY for mock server
 - Turn on tests in AppVeyor build

+++ test/ares-test-misc.cc
@@ -239,7 +239,7 @@ TEST_F(DefaultChannelTest, SendFailure) {
   EXPECT_EQ(ARES_EBADQUERY, result.status_);
 }
 
-std::string ExpandName(const std::vector<byte>& data, ssize_t offset,
+std::string ExpandName(const std::vector<byte>& data, int offset,
                        long *enclen) {
   char *name = nullptr;
   int rc = ares_expand_name(data.data() + offset, data.data(), data.size(),
