+++ test/ares-test.h
@@ -64,7 +64,7 @@ class LibraryTest : public ::testing::Test {
   static void afree(void *ptr);
  private:
   static bool ShouldAllocFail(size_t size);
-  static unsigned long fails_;
+  static unsigned long long fails_;
   static std::map<size_t, int> size_fails_;
 };
